pyproject.toml
[project]
name = "nonebot-plugin-multincm"
dynamic = ["version"]
description = "NCM Song Searcher"
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.4.0",
    "nonebot-plugin-htmlrender>=0.5.1",
    "nonebot-plugin-alconna>=0.54.1",
    "pyncm>=1.7.1",
    "typing-extensions>=4.12.2",
    "httpx>=0.27.2",
    "jinja2>=3.1.5",
    "cookit[jinja,loguru,nonebot-alconna,pydantic]>=0.8.4",
    "nonebot-plugin-waiter>=0.8.0",
    "cachetools>=5.5.0",
    "yarl>=1.18.3",
    "silk-python>=0.2.6",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-multincm"

[dependency-groups]
dev = ["nonebot-adapter-kritor>=0.2.2", "nonebot-adapter-onebot>=2.4.3"]

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_multincm/__init__.py"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
